refactor with components
